Refactor tests x2
