--- tests/Fixtures/ExampleA/FacadeInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace GacelaTest\Fixtures\ExampleA;
-
-interface FacadeInterface
-{
-    public function greet(string $name): array;
-}
