--- tests/Fixtures/ExampleA/Service/ServiceA.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace GacelaTest\Fixtures\ExampleA\Service;
-
-final class ServiceA
-{
-    public function greet(string $name): array
-    {
-        return ["Hello, $name from A."];
-    }
-}
