--- tests/Fixtures/ExampleB/FacadeInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace GacelaTest\Fixtures\ExampleB;
-
-interface FacadeInterface
-{
-    public function greet(string $name): array;
-}
