--- tests/Fixtures/ExampleC/Facade.php
@@ -1,20 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace GacelaTest\Fixtures\ExampleC;
-
-use Gacela\AbstractFacade;
-
-/**
- * @method Factory getFactory()
- */
-final class Facade extends AbstractFacade implements FacadeInterface
-{
-    public function greet(string $name): array
-    {
-        return $this->getFactory()
-            ->createGreeter()
-            ->greet($name);
-    }
-}
