--- tests/Fixtures/ExampleC/FacadeInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace GacelaTest\Fixtures\ExampleC;
-
-interface FacadeInterface
-{
-    public function greet(string $name): array;
-}
