--- tests/Fixtures/MissingConfigModule/MissingConfigModuleFactory.php
@@ -1,15 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace GacelaTest\Fixtures\MissingConfigModule;
-
-use Gacela\AbstractFactory;
-
-final class MissingConfigModuleFactory extends AbstractFactory
-{
-    public function createDomainService(): void
-    {
-        $this->getConfig();
-    }
-}
