+++ tests/Integration/RemoveKeyFromContainer/AddAndRemoveKey/DependencyProvider.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace GacelaTest\Integration\MissingFile\RemoveKeyFromContainer;
+namespace GacelaTest\Integration\RemoveKeyFromContainer\AddAndRemoveKey;
 
 use Gacela\AbstractDependencyProvider;
 use Gacela\Container\Container;
