+++ tests/Integration/RemoveKeyFromContainer/AddAndRemoveKey/Factory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace GacelaTest\Integration\MissingFile\RemoveKeyFromContainer;
+namespace GacelaTest\Integration\RemoveKeyFromContainer\AddAndRemoveKey;
 
 use Gacela\AbstractFactory;
