+++ tests/Integration/UsingConfig/config.php
@@ -2,4 +2,4 @@
 
 declare(strict_types=1);
 
-$config['test-number'] = 1;
+$config['number'] = 1;
